convert SubmissionsAdjustmentTest
